[COMMUNITY] Enable Linux-based hypervisor execution in AGL
The upstream version of AGL boots Linux in EL1. This mode is not
privileged enough to put in place the correct configuration of
an hypervisor (e.g., KVM).

This patch leverages the ATF rcar platform configuration variables to
boot the Linux kernel in EL2, instead of EL1, allowing Linux-based
hypervisors such as KVM to configure all the necessary stubs to
properly run virtual machines.

This patch has been successfully tested with ATF, KVM and AGL rootfs
on the Renesas R-Car M3.

Changelog:
- Added arm-trusted-firmware patch, which makes this work easy
  to be tested by the community
- This patch version does not contain Linux kernel modifications.
  In fact, these resulted not necessary with the proposed ATF patch
  and a recent kernel (higher than 4.6)

Bug-AGL: SPEC-497

Change-Id: I5a22411135b33137cb2236562b3de9c966f1be21
Signed-off-by: Michele Paolino <m.paolino@virtualopensystems.com>