gstreamer1.0-plugin-vspfilter: add missing inherit pkgconfig
pkg-config is called in the do_configure() task. With the introduction
pf per recipe sysroot add the missing pkgconfig to the inherit variable

Fixes configure task error of the form:
configure:12146: checking for GLIB
configure:12214: result: no
configure:12216: error: in `gstreamer1.0-plugin-vspfilter/1.0.0-r0/build':
configure:12218: error: The pkg-config script could not be found or is too old. Make sure it
is in your PATH or set the PKG_CONFIG environment variable to the full path to pkg-config.

Signed-off-by: Stephen Lawrence <stephen.lawrence@renesas.com>