[COMMUNITY] ADSP enable and add sound hardware abstraction in DT
The ADSP on Renesas SoCs required a reserved memory area to become
enabled, and this needs to be done manually.

This patch adds sound hardware abstraction information for the m3ulcb
and m3ulcb-kf (kingfisher) device trees. This is helpful on-board the
ADSP for dynamically determining sound hardware at runtime, allowing
single binaries to be used between the two boards.

Future work will incorporate the h3ulcb and Salvator boards into the
abstraction.

Bug-AGL: SPEC-1424

Change-Id: Ie3f3d2227e79a6396f056fc63ecbcc6c73034e6b
Signed-off-by: Mark Farrugia <mark.farrugia@fiberdyne.com.au>