[COMMUNITY] Fix SSI DMAC inaccessible via Xtensa HiFi2 ADSP firmware
The upgrade of the kernel to 4.14.35 has changed the SSI interface in
the RCar SoC sound driver, making it that the Xtensa HiFi2 ADSP firmware
can no longer directly access the SSI DMA controller in order to
route audio directly out to hardware. This seems to fix this issue,
allowing the Fiberdyne DSP to fully operate again on 4.14.35 kernels.

Change-Id: Id89843b0a8a65f07964027639dde68791261cab6
Signed-off-by: Mark Farrugia <mark.farrugia@fiberdyne.com.au>