[AGL] Apply the syscall restarting patch for aarch64 SoC
There are some seccomp bpf related failures when running kselftest
on R-Car M3(AGL inside).

Now, apply this patch to fix the syscall restart related bug(only
existing on kernel version < 4.14.53).

This patch won't be necessary if the kernel that AGL using was
changed to the latest ltsi(4.14.75) or kernel >= 4.14.53 in the future.

Change-Id: Ifeb741b16ef5d4d76c7c86419ddc7817a5a3821b
Signed-off-by: Liu Wenlong <liuwl.fnst@cn.fujitsu.com>